Day 6 code: RecyclerView
Added a new module for Day 6 focusing on RecyclerView implementation.
Created a layout for student items and an adapter to display student data in a RecyclerView.
The MainActivity now uses this adapter and layout to show a list of students.
Clicking on a student item triggers a Toast message displaying the student's name.